refactor: update to new `inspect` dir structure
